wip: Add customizable service accounts to worker nodes
According to the docs, there is a minimum permission needed for worker nodes to function but
my attempt to add that fails with a message about `roles/container.defaultNodeServiceAccount`
not being able to be assigned to the service account.

It seems like the worker nodes are able to function without this permission though, so I'm
going to leave it as is for now.

See: https://cloud.google.com/kubernetes-engine/docs/how-to/creating-a-regional-cluster#setup-iam-service-account